build fixes
protobuf: in line 44 cmake snippets are included from
${PROTOBUF_INSTALL_DIR}/lib/cmake/protobuf, so it should be made sure
that they are actually installed in lib (and not lib64), therefore add
CMAKE_INSTALL_LIBDIR=lib.

hef_proto, scheduler_mon_proto and profiler_proto are private libraries
that are not meant to be installed but are used by libhailort, so make
sure they are built as static libraries.